refactor: 클래스 네이밍을 구체적으로 수정
