+++ app/src/main/java/woowacourse/movie/list/view/MovieListFragment.kt
@@ -9,16 +9,16 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import woowacourse.movie.R
 import woowacourse.movie.databinding.FragmentHomeBinding
-import woowacourse.movie.list.adapter.MovieAdapter
+import woowacourse.movie.list.adapter.MovieListAdapter
 import woowacourse.movie.list.contract.MovieListContract
 import woowacourse.movie.list.model.Advertisement
 import woowacourse.movie.list.model.Movie
 import woowacourse.movie.list.presenter.MovieListPresenter
 
-class HomeFragment : Fragment(), MovieListContract.View {
+class MovieListFragment : Fragment(), MovieListContract.View {
     override val presenter = MovieListPresenter(this)
     private lateinit var binding: FragmentHomeBinding
-    private lateinit var movieAdapter: MovieAdapter
+    private lateinit var movieListAdapter: MovieListAdapter
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -38,14 +38,14 @@ class HomeFragment : Fragment(), MovieListContract.View {
         movies: List<Movie>,
         advertisements: List<Advertisement>,
     ) {
-        movieAdapter = MovieAdapter(movies, advertisements)
-        binding.movieRecyclerView.adapter = movieAdapter
+        movieListAdapter = MovieListAdapter(movies, advertisements)
+        binding.movieRecyclerView.adapter = movieListAdapter
         binding.movieRecyclerView.layoutManager = LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
     }
 
     override fun setOnListViewClickListener() {
-        movieAdapter.setItemClickListener(
-            object : MovieAdapter.OnItemClickListener {
+        movieListAdapter.setItemClickListener(
+            object : MovieListAdapter.OnItemClickListener {
                 override fun onClick(movieId: Long) {
                     val theaterBottomSheetFragment = TheaterBottomSheetFragment()
                     val bundle = Bundle()
