+++ app/src/main/res/layout/fragment_home.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="fragment"
-            type="woowacourse.movie.list.view.HomeFragment" />
+            type="woowacourse.movie.list.view.MovieListFragment" />
 
         <variable
             name="presenter"
@@ -18,7 +18,7 @@
         android:orientation="vertical"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".list.view.HomeFragment">
+        tools:context=".list.view.MovieListFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/movie_recycler_view"
